reorganize server code into directory
